rename internal package, replace forked module, fix vet
