+++ assert_http.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/beme/abide/internal"
+	"github.com/beme/abide/utils"
 	"github.com/nsf/jsondiff"
 	"github.com/sergi/go-diff/diffmatchpatch"
 )
@@ -77,7 +77,7 @@ func (h *httpRequest) jsonBodyCleanup(t *testing.T, c *config) error {
 	// Clean/update json based on config.
 	if c != nil {
 		for k, v := range c.Defaults {
-			jsonIface = internal.UpdateKeyValuesInMap(k, v, jsonIface)
+			jsonIface = utils.UpdateKeyValuesInMap(k, v, jsonIface)
 		}
 	}
