+++ go.mod
@@ -7,3 +7,5 @@ require (
 	github.com/nsf/jsondiff v0.0.0-20210926074059-1e845ec5d249
 	github.com/sergi/go-diff v1.2.0
 )
+
+replace github.com/beme/abide => ./
