+++ utils/json.go
@@ -1,4 +1,4 @@
-package internal
+package utils
 
 // UpdateKeyValuesInMap updates every instance of a key within an arbitrary
 // `map[string]interface{}` with the given value.
